Save fs register before making an OCall
An OCall consists of an EEXIT followed by an EENTER upon return. On
EENTER, SGX will restore fs and gs values stored in the TCS structure
for that thread. If the enclave application has modified the fs
register, SGX will not restore it.

Modify oe_ocall to save this value if it has changed prior to making an
OCall and restore it upon return.

Signed-off-by: Jordan Hand <jorhand@microsoft.com>